+++ types/userData.ts
@@ -1,0 +1,8 @@
+export type UserData = {
+  email: string
+  bio: string
+  profileImage: string
+  nickname: string
+  friends?: string[]
+  friendsDetails?: { id: string; nickname: string; bio: string }[]
+}
